refactor(auth): extract platform checks into auth-platform-utils
